bilder verschoben und floats durch ints ersetzt in aufgabe2

+++ Uebung1/Uebung1_Aufgab2_1.py
@@ -1,6 +1,6 @@
-zahl1 = float(input("Zahl1:"))
-zahl2 = float(input("Zahl2:"))
-zahl3 = float(input("Zahl3:"))
+zahl1 = int(input("Zahl1:"))
+zahl2 = int(input("Zahl2:"))
+zahl3 = int(input("Zahl3:"))
 
 Summe = zahl1 + zahl2 + zahl3   #summe der 3 Inputs bilden
